requirements.txt
alabaster==0.7.12
Babel==2.8.0
breathe==4.19.2
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
docutils==0.16
et-xmlfile==1.0.1
future==0.18.2
idna==2.10
imagesize==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
junitparser==1.4.1
kiwisolver==1.2.0
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mlx.traceability==4.6.1
natsort==7.0.1
numpy==1.19.0
openpyxl==3.0.4
packaging==20.4
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.1.2
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
urllib3==1.25.9
